requirements.txt
streamlit
numpy
pillow
streamlit-drawable-canvas